slight refactoring

+++ core/include/lexicon.hpp
@@ -9,7 +9,7 @@
 class Lexicon {
 public:
     // Adds a word and returns its id
-    uint32_t add_word(std::string word);
+    uint64_t add_word(std::string word);
 
     // Batch word addition, input is expected to be normalized
     void add_words(std::vector<std::string> words);
@@ -18,9 +18,9 @@ public:
     std::string get_word(int word_id);
 
     // Returns 0 if no id exists
-    uint32_t get_word_id(std::string word);
+    uint64_t get_word_id(std::string word);
 
-    uint32_t size();
+    uint64_t size();
 
     // Converts a given chunk of text into normalized tokens
     static std::vector<std::string> tokenize_text(std::string text, char delim = ' ');
@@ -39,14 +39,12 @@ public:
 
 private:
     std::vector<std::string> id_to_word;
-    std::unordered_map<std::string, uint32_t> word_to_id;
+    std::unordered_map<std::string, uint64_t> word_to_id;
 
-    uint32_t next_id = 1; // 0 is for not found
+    uint64_t next_id = 1; // 0 is for not found
 
 
-
-    // Removes leading and lagging spaces
-    // TODO: Make a separate file for such utility functions
+    // Removes leading and lagging spaces + any other specified characters
     static std::string trim(const std::string & source, std::string remove_left = "", std::string remove_right = "");
 };
